virtual-try-on/requirements.txt
google-cloud-aiplatform==1.36.4
google-cloud-storage==2.5.0
img2texture==1.1.2.post2
imutils==0.5.4
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.1
rembg==2.0.52
Pillow==10.1.0
PyYAML==6.0.1
vertexai==0.0.1
opencv-contrib-python==4.8.1.78
gradio==4.9.1